+++ src/app/components/tickets/tickets.component.ts
@@ -3,6 +3,7 @@ import {
   ChangeDetectionStrategy,
   Component,
   computed,
+  effect,
   inject,
   OnDestroy,
   OnInit,
@@ -18,11 +19,16 @@ import {
 } from '@models/product.model';
 import { MappingService } from '@services/mapping.service';
 import { WcStoreAPI } from '@services/api/wc-store-api.service';
-import { Subject, switchMap, takeUntil } from 'rxjs';
+import { catchError, Subject, switchMap, takeUntil, throwError } from 'rxjs';
 import { ProductComponent } from './product/product.component';
 import { LocalStorageKeys } from '@models/storage.model';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { crossSaleProductCat } from '@models/cross-sale.model';
+import { ErrorDialogService } from '@shared/errors/error-dialog.service';
+import { DisclaimerComponent } from './disclaimer/disclaimer.component';
+import { DisclaimerForm } from '@models/disclaimer.model';
+import { DisclaimerStateService } from '@services/disclaimer-state.service';
 
 @Component({
   selector: 'app-tickets',
@@ -34,6 +40,7 @@ import { MatProgressBarModule } from '@angular/material/progress-bar';
     MatProgressSpinnerModule,
     MatProgressBarModule,
     CommonModule,
+    DisclaimerComponent,
   ],
   templateUrl: './tickets.component.html',
   styleUrl: './tickets.component.scss',
@@ -41,9 +48,11 @@ import { MatProgressBarModule } from '@angular/material/progress-bar';
 })
 export class TicketsComponent implements OnInit, OnDestroy, AfterViewInit {
   private readonly destroy$ = new Subject<void>();
-  mappingService = inject(MappingService);
-  wcStore = inject(WcStoreAPI);
-  listVariations = ['instock']; // add 'outofstock' to show out-of-stock stuff
+  private readonly mappingService = inject(MappingService);
+  private readonly errorService = inject(ErrorDialogService);
+  private readonly disclaimerStateS = inject(DisclaimerStateService);
+  private readonly wcStore = inject(WcStoreAPI);
+  listVariations = ['instock']; // add 'outofstock' here, to show out-of-stock variations
 
   /** loading */
   viewLoading = true;
@@ -51,12 +60,40 @@ export class TicketsComponent implements OnInit, OnDestroy, AfterViewInit {
 
   /** signals */
   products = signal<Array<WcProduct>>([]);
+  crossSaleProducts = signal<Array<WcProduct>>([]);
   selectedProduct = signal<WcProduct | undefined>(undefined);
   isSelectedProductVariable = signal<boolean>(false);
   selectedProductVariations = signal<Array<WcProductVariationDetails> | []>([]);
   selectedProductId = computed(() => {
     return this.selectedProduct()?.id;
   });
+  selectedProductHasDisclaimer = computed(() => {
+    return this.disclaimerStateS.hasProductDisclaimer(
+      this.selectedProduct()?.id,
+    );
+  });
+  /** returns true, if the selectedProduct is marked as having a disclaimer and its valid */
+  selectedProductValidDisclaimer = computed(() => {
+    if (this.selectedProductHasDisclaimer()) {
+      return this.disclaimerStateS.validateDisclaimerState(
+        this.selectedProduct()!.id,
+      );
+    } else return false;
+  });
+  /** on signal change, will set the disclaimer for the product */
+  setDisclaimer = effect(() => {
+    const product = this.selectedProduct();
+    if (this.selectedProductHasDisclaimer()) {
+      this.disclaimerStateS.setDisclaimer(product!);
+    }
+  });
+
+  // TODO
+  /**
+   * The cleanest approach here would probably be to define a parent FC and having the mat-cards as main-controls.
+   * Otherwise, since we query the product on select anyways (which could also be avoided probably),
+   * we might want to destroy the productDetails on select
+   */
 
   constructor() {}
 
@@ -66,33 +103,62 @@ export class TicketsComponent implements OnInit, OnDestroy, AfterViewInit {
 
   ngOnInit(): void {
     this.initProducts();
+    this.queryCrossSaleOptions();
 
     const storedSelect = this.getStoredSelect();
     // query the product and set it as selected
     if (storedSelect) this.querySelectedProduct(storedSelect);
   }
 
-  querySelectedProduct(id: number) {
+  private queryCrossSaleOptions() {
+    this.productsLoading.set(true);
+    this.wcStore
+      .listProducts(crossSaleProductCat)
+      .pipe(
+        catchError((error) => {
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
+        takeUntil(this.destroy$),
+      )
+      .subscribe((r) => {
+        this.crossSaleProducts.set(r);
+        this.productsLoading.set(false);
+      });
+  }
+
+  private querySelectedProduct(id: number) {
     this.productsLoading.set(true);
     this.wcStore
       .getProductById(id)
       .pipe(
         takeUntil(this.destroy$),
+        catchError((error) => {
+          this.productsLoading.set(false);
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
         switchMap((product) => {
           this.selectedProduct.set(product);
           this.productsLoading.set(false);
-          this.isSelectedProductVariable.set(
-            product.type === WcProductTypes.VARIABLE,
-          );
 
           if (product.type === WcProductTypes.VARIABLE) {
+            this.isSelectedProductVariable.set(true);
             this.productsLoading.set(true);
             return this.wcStore.listProductVariations(id, this.listVariations);
           } else {
+            this.isSelectedProductVariable.set(false);
             return []; // Return an empty array if it's not a variable product
           }
         }),
         takeUntil(this.destroy$),
+      )
+      .pipe(
+        catchError((error) => {
+          this.productsLoading.set(false);
+          this.errorService.handleError(error);
+          return throwError(() => error);
+        }),
       )
       .subscribe((response) => {
         this.selectedProductVariations.set(response.length > 0 ? response : []);
@@ -127,6 +193,21 @@ export class TicketsComponent implements OnInit, OnDestroy, AfterViewInit {
     } else {
       id ? this.querySelectedProduct(id) : new Error();
       localStorage.setItem(LocalStorageKeys.PRODUCT_SELECTED_ID, id.toString());
+    }
+  }
+
+  onCloseDisclaimer(event: boolean) {
+    // unselect the product on close because disclaimer invalid
+    this.selectedProduct.set(undefined);
+  }
+
+  onDisclaimerSubmitted(event: DisclaimerForm) {
+    this.productsLoading.set(true);
+    const currentSelected = this.selectedProductId();
+    if (currentSelected) {
+      this.disclaimerStateS.pushDisclaimerState(currentSelected, event);
+      // trigger signal chain to destroy disclaimer if valid disclaimer state, should also kill the loading animation
+      this.onProductSelected(currentSelected);
     }
   }