+++ src/app/models/cross-sale.model.ts
@@ -1,0 +1,7 @@
+export enum CrossSaleProductId {
+  SOLI = 1960,
+  GOENNER = 1961,
+  KONFUSIUS = 1962,
+}
+
+export const crossSaleProductCat = 49;