+++ src/app/utils/price.utils.ts
@@ -1,9 +1,12 @@
+import { WcPriceRange } from '@models/price.model';
+
 export function formatPrice(
   price: string,
   thSep: string,
   decSep: string,
   cmu: number,
-) {
+  currency?: string,
+): string {
   let numericTotalPrice = parseFloat(price ?? 'NaN');
 
   const factor = Math.pow(10, cmu ?? 2);
@@ -13,8 +16,20 @@ export function formatPrice(
 
     let [intP, decP] = fPrice.split('.');
     intP = intP.replace(/\B(?=(\d{3})+(?!\d))/g, thSep);
-    return `${intP}${decSep}${decP}`;
+    return `${intP}${decSep}${decP} ${currency}`;
   };
 
   return format(numericTotalPrice);
+}
+
+export function formatPriceRange(
+  priceRange: WcPriceRange,
+  thSep: string,
+  decSep: string,
+  cmu: number,
+  currency?: string,
+): string {
+  const min = formatPrice(priceRange.min_amount, thSep, decSep, cmu, currency);
+  const max = formatPrice(priceRange.max_amount, thSep, decSep, cmu, currency);
+  return `${min} bis ${max}`;
 }